Added first two Brainfuck instructions.
Currently the interpreter processes only two instructions - '+' and
'-'. Other characters are stripped out during the 'compile' phase.

The Brainfuck program is hardcoded as +++++[]--- and is passed in
to the BFInterpreter_compile procedure. As mentioned above, the
non-recognised characters, in this case [] are stripped when the
program is 'written to' the 'program tape'. Thus the program that
is evaluated when the BFInterpreter_eval routine is called is
+++++---

By using the DEBUG flag during compilation, debug messages are
printed out.