Added further instructions and some basic error checking.
The interpreter now recognises the instructions - '+' '-' '.' ','
'>' and '<'. The loop instructions ']' and '[' are not yet
recognised. The program is still hardcoded.

Basic error checking is in place to ensure that the program_tape
array does not overflow. However the error checking does need to
be made more robust...

Next steps will be to add a simple stack datatype to handle the
two loop instructions. The error handling will need to be made
more robust too.