Added the loop instructions.
Changed the representation of a Brainfuck instruction from a simple
char to a tagged union - either a simple instruction like + or a
loop instruction like [. This allows a simple way of keeping track
of the jump locations where needed.

Hardcoded a Hello World program for testing - worked correctly.
Extra error checking will need to be put in place.