Added error checking and ability to read in code from source files.
The representation of instructions is a little messy - mainly the
need for the enumerated type. Probably should look in to tidying
that up.

When compiling with -O3 flags, this interpreter computes the first
100 primes in about 0.6 - 0.7 seconds. Executing the interpreter
with the mandelbrot.b source file takes about 48 seconds. Timings
were taken on a 2012 Macbook Air with 8GB of RAM and a 2GHz i7
processor. So looks like there is room for improvement.